fix: restore site live preview after channel migration
The site's preview never set `window.channelBridge`, so every widget that
moved onto channels (#676, #678, #685) threw `Cannot read properties of
undefined (reading 'subscribe')` on mount. `SectorTimingUpdater` sits outside
any error boundary, so that throw unmounted the whole page and irdashies.com
rendered blank.

Run the real pipeline in the browser rather than shipping canned snapshots,
so the preview stays animated:

- Extract `ChannelBus` into `channelBus.ts`. The class was already free of
  Electron; only `targetFor`/`setupChannelBridge` need it. `channelBridge.ts`
  keeps those and re-exports, so main-process imports are unchanged.
- Drop the static electron-log import from `ProcessorHost`. `logError` was
  already injectable, so it now defaults to a no-op and the two Electron
  composition points pass the real logger explicitly.
- Add `previewChannelRuntime`, which runs the same bus and processors in-page
  against the mock SDK source and hands widgets a `ChannelBridge`. Demand
  aggregation mirrors `channelRendererBridge`; the virtual renderer target
  tracks tab visibility so a backgrounded tab costs nothing.
- Wrap the headless updaters in an error boundary so a channel fault degrades
  one widget instead of blanking the page, and log what it caught.
