+++ site/src/components/WidgetErrorBoundary.tsx
@@ -1,8 +1,15 @@
 import { Component, type ReactNode } from 'react';
+import logger from '@irdashies/utils/logger';
 
 interface Props {
   children: ReactNode;
   widgetName: string;
+  /**
+   * Rendered instead of the default placeholder when the subtree throws. Pass
+   * `null` for headless children, where a visible placeholder would inject a
+   * stray box into the layout.
+   */
+  fallback?: ReactNode;
 }
 
 interface State {
@@ -16,8 +23,13 @@ export class WidgetErrorBoundary extends Component<Props, State> {
     return { hasError: true };
   }
 
+  componentDidCatch(error: unknown) {
+    logger.error(`[preview] ${this.props.widgetName} failed to render`, error);
+  }
+
   render() {
     if (this.state.hasError) {
+      if (this.props.fallback !== undefined) return this.props.fallback;
       return (
         <div className="flex items-center justify-center h-full bg-slate-900/60 rounded-sm border border-slate-800/50 p-4">
           <p className="text-xs text-slate-500 uppercase tracking-wide font-bold">
